ログイン周りのコード整理
