+++ app/Http/Controllers/Admin/HomeController.php
@@ -3,17 +3,15 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\BaseController;
-use Illuminate\Contracts\Support\Renderable;
+use Illuminate\View\View;
 
 class HomeController extends BaseController
 {
 
     /**
      * ホーム画面の初期表示
-     *
-     * @return Renderable
      */
-    public function index(): Renderable
+    public function index(): View
     {
         return view('admin.home');
     }
