+++ resources/views/admin/contact/edit.blade.php
@@ -1,4 +1,4 @@
-@extends('layouts.app_admin')
+@extends('layouts.admin')
 @section('title',  __('order.Order ID:') . $contactForm->id. __('common.Of Change'))
 @php
     $menu = 'user';
