+++ resources/views/admin/contact/show.blade.php
@@ -1,4 +1,4 @@
-@extends('layouts.app_admin')
+@extends('layouts.admin')
 @section('title', __('contact.Contact ID:') . $contactForm->id)
 @php
     $menu = 'user';
