+++ resources/views/admin/home.blade.php
@@ -1,4 +1,4 @@
-@extends('layouts.app_admin')
+@extends('layouts.admin')
 @section('title', __('common.HOME'))
 
 @section('breadcrumbs')
