+++ resources/views/admin/order/index.blade.php
@@ -1,4 +1,4 @@
-@extends('layouts.app_admin')
+@extends('layouts.admin')
 @section('title', __('order.Order List'))
 @php
     $menu = 'master';
