+++ resources/views/admin/order/show.blade.php
@@ -1,4 +1,4 @@
-@extends('layouts.app_admin')
+@extends('layouts.admin')
 @section('title', __('order.Order ID:') . $order->id)
 @php
     $menu = 'master';
