+++ resources/views/admin/passwordChange.blade.php
@@ -1,4 +1,4 @@
-@extends('layouts.app_admin')
+@extends('layouts.admin')
 @section('title', __('common.Password Change'))
 
 @section('breadcrumbs')
