+++ resources/views/admin/staff/edit.blade.php
@@ -1,4 +1,4 @@
-@extends('layouts.app_admin')
+@extends('layouts.admin')
 @section('title', $staff->name . __('common.Of Change'))
 @php
     $menu = 'system';
