+++ resources/views/admin/staff/import.blade.php
@@ -1,4 +1,4 @@
-@extends('layouts.app_admin')
+@extends('layouts.admin')
 @section('title',  __('staff.Staff Import'))
 @php
     $menu = 'system';
