+++ resources/views/admin/stock/create.blade.php
@@ -1,4 +1,4 @@
-@extends('layouts.app_admin')
+@extends('layouts.admin')
 @section('title', __('stock.Stock Regist'))
 @php
     $menu = 'master';
