+++ resources/views/admin/stock/index.blade.php
@@ -1,4 +1,4 @@
-@extends('layouts.app_admin')
+@extends('layouts.admin')
 @section('title', __('stock.Stock List'))
 @php
     $menu = 'master';
