+++ resources/views/admin/stock/show.blade.php
@@ -1,4 +1,4 @@
-@extends('layouts.app_admin')
+@extends('layouts.admin')
 @section('title', $stock->name)
 @php
     $menu = 'master';
