+++ resources/views/admin/user/show.blade.php
@@ -1,4 +1,4 @@
-@extends('layouts.app_admin')
+@extends('layouts.admin')
 @section('title', $user->name)
 @php
     $menu = 'user';
