+++ resources/views/layouts/front.blade.php
